:recycle: refactor: 체형 진단 결과 페이지 컴포넌트 분리
